+++ VOS_PRJ/vos_test/delay_test.c
@@ -8,7 +8,7 @@ static void task0(void *param)
 {
 	int cnts = 0;
 	kprintf("%s start ...\r\n", __FUNCTION__);
-	s64 mark_time = VOSGetTimeMs()/1000;
+	u32 mark_time = VOSGetTimeMs()/1000;
 	kprintf("mark_time: %d\r\n", (u32)mark_time);
 #if 1
 	while (TestExitFlagGet() == 0) {
