+++ VOS_PRJ/vos_test/mq_test.c
@@ -15,7 +15,7 @@ static void task0(void *param)
 	int cnts = 0;
 	StMsgBlock tmp;
 	kprintf("%s start ...\r\n", __FUNCTION__);
-	s64 mark_time = VOSGetTimeMs()/1000;
+	u32 mark_time = VOSGetTimeMs()/1000;
 	kprintf("mark_time: %d\r\n", (u32)mark_time);
 
 	gMsgQuePtr = VOSMsgQueCreate(arr_msg, sizeof(arr_msg), sizeof(StMsgBlock), "msg test");
